put all routes in routes file and be imported in __init__.py
