+++ taipei-day-trip/TaipeiTravel/__init__.py
@@ -6,3 +6,5 @@ app=Flask(__name__)
 app.config["JSON_AS_ASCII"]=False
 app.json.ensure_ascii = False
 app.config["TEMPLATES_AUTO_RELOAD"]=True
+
+from TaipeiTravel import routes, attraction, member, payment, itinerary
