+++ taipei-day-trip/TaipeiTravel/routes.py
@@ -1,0 +1,22 @@
+from TaipeiTravel import app
+from flask import render_template
+
+# Pages
+@app.route("/")
+def index():
+	return render_template("index.html")
+
+
+@app.route("/attraction/<id>")
+def attraction(id):
+	return render_template("attraction.html")
+
+
+@app.route("/booking")
+def booking():
+	return render_template("booking.html")
+
+
+@app.route("/thankyou")
+def thankyou():
+	return render_template("thankyou.html")
